# Conflicts:
#	app/src/main/res/layout/activity_new_trip.xml

+++ app/src/main/java/de/traveltogether/invitation/IInviteInteractor.java
@@ -7,5 +7,5 @@ import de.traveltogether.IInteractor;
  */
 public interface IInviteInteractor extends IInteractor {
     public void getFormerParticipants(IInvitePresenter _presenter);
-    public void invite(int tripId, long personId, IInvitePresenter listener);
+    public void invite(long tripId, int personId, IInvitePresenter listener);
 }
